backend/requirements.txt
Django>=4.2
django-cors-headers>=4.3.1
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1
django-extensions>=3.2.3
django-filter~=25.1
pip~=24.3.1
attrs~=25.3.0
asgiref~=3.8.1
sqlparse~=0.5.3
PyYAML~=6.0.2
Jinja2~=3.1.6
requests~=2.32.3
coreapi~=2.3.3
uritemplate~=4.1.1
coreschema~=0.0.4
inflection~=0.5.1
PyJWT~=2.9.0
PyMySQL~=1.1.1
cryptography~=44.0.2
pillow~=11.1.0
cffi~=1.17.1
typing_extensions~=4.13.1
MarkupSafe~=3.0.2
certifi~=2025.1.31
itypes~=1.2.0
urllib3~=2.3.0
idna~=3.10
charset-normalizer~=3.4.1
simplejson~=3.20.1
pycparser~=2.22
